requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
Pillow>=9.0.0
matplotlib>=3.7.0
tqdm>=4.60.0
clip @ git+https://github.com/openai/CLIP.git
transformers>=4.30.0
pycocotools>=2.0.6
